+++ app/src/main/java/io/jackson/instacopy/ViewEffects.kt
@@ -1,5 +1,6 @@
-package io.jackson.instacopy.store
+package io.jackson.instacopy
 
 
 sealed class ViewEffect
+
 data class ShowPickerViewEffect(val itemId: String): ViewEffect()
