+++ app/src/main/java/io/jackson/instacopy/repo/MockRepository.kt
@@ -128,5 +128,12 @@ object MockRepository : StoreRepository {
 
 val testItemCarouselPlaceholder = ItemCarouselViewModel(
         title = "Based on your cart",
-        items = listOf(ItemViewModel.PLACE_HOLDER, ItemViewModel.PLACE_HOLDER, ItemViewModel.PLACE_HOLDER, ItemViewModel.PLACE_HOLDER, ItemViewModel.PLACE_HOLDER, ItemViewModel.PLACE_HOLDER, ItemViewModel.PLACE_HOLDER, ItemViewModel.PLACE_HOLDER)
+        items = listOf(ItemViewModel.PLACE_HOLDER,
+                ItemViewModel.PLACE_HOLDER,
+                ItemViewModel.PLACE_HOLDER,
+                ItemViewModel.PLACE_HOLDER,
+                ItemViewModel.PLACE_HOLDER,
+                ItemViewModel.PLACE_HOLDER,
+                ItemViewModel.PLACE_HOLDER,
+                ItemViewModel.PLACE_HOLDER)
 )
