+++ app/src/main/java/io/jackson/instacopy/store/StoreFragment.kt
@@ -3,6 +3,7 @@ package io.jackson.instacopy.store
 import android.os.Bundle
 import android.os.Handler
 import android.util.Log
+import android.view.ContextMenu
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -13,6 +14,7 @@ import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import androidx.recyclerview.widget.SimpleItemAnimator
 import com.beyondeye.reduks.StoreSubscriber
+import com.beyondeye.reduks.StoreSubscription
 import io.jackson.instacopy.*
 import io.jackson.instacopy.boundary.toViewModels
 import io.jackson.instacopy.middleware.ViewEffectsMiddleware
@@ -34,6 +36,7 @@ class StoreFragment : Fragment(), CoroutineScope, StoreSubscriber<AppState> {
     private lateinit var storeAdapter: StoreRecyclerViewAdapter
     private val loc = IntArray(2)
     private val repo = RetrofitStoreRepository
+    private var subscription: StoreSubscription? = null
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
         return inflater.inflate(R.layout.content_main, container, false)
@@ -47,7 +50,7 @@ class StoreFragment : Fragment(), CoroutineScope, StoreSubscriber<AppState> {
             adapter = storeAdapter
         }
 
-        val subscription = appStore.subscribe(this)
+        subscription = appStore.subscribe(this)
 
         appStore.dispatch(NetworkThunks.fetchStoreInfoThunk("sprouts"))
 
@@ -67,6 +70,7 @@ class StoreFragment : Fragment(), CoroutineScope, StoreSubscriber<AppState> {
             appStore.dispatch(NetworkThunks.fetchCoupons("sprouts"))
         }, 8000)
 
+        (rootRecyclerView.itemAnimator as SimpleItemAnimator).supportsChangeAnimations = false
         rootRecyclerView.addOnScrollListener(
                 object : RecyclerView.OnScrollListener() {
                     override fun onScrolled(recyclerView: RecyclerView, dx: Int, dy: Int) {
@@ -91,20 +95,22 @@ class StoreFragment : Fragment(), CoroutineScope, StoreSubscriber<AppState> {
 
     }
 
-    override fun onStateChange() {
-        activity?.runOnUiThread {
-            (rootRecyclerView.itemAnimator as SimpleItemAnimator).supportsChangeAnimations = false
-            storeAdapter.setListData(appStore.state.listData.toViewModels(appStore.state.cart).toMutableList())
-            if (appStore.state.cart.totalNumItems() > 0) {
-                activity?.findViewById<TextView>(R.id.btnCartQuantity)!!.visibility = View.VISIBLE
 
-                activity?.btnCartQuantity!!.text = appStore.state.cart.totalNumItems().toString()
-                activity?.btnTopCartQuantity!!.visibility = View.VISIBLE
-                activity?.btnTopCartQuantity!!.text = appStore.state.cart.totalNumItems().toString()
-            } else {
-                activity?.btnCartQuantity!!.visibility = View.GONE
-                activity?.btnTopCartQuantity!!.visibility = View.GONE
-            }
+    override fun onDestroyView() {
+        subscription?.unsubscribe()
+        super.onDestroyView()
+    }
+
+    override fun onStateChange() {
+        storeAdapter.setListData(appStore.state.listData.toViewModels(appStore.state.cart).toMutableList())
+        if (appStore.state.cart.totalNumItems() > 0) {
+            activity?.btnCartQuantity!!.visibility = View.VISIBLE
+            activity?.btnCartQuantity!!.text = appStore.state.cart.totalNumItems().toString()
+            activity?.btnTopCartQuantity!!.visibility = View.VISIBLE
+            activity?.btnTopCartQuantity!!.text = appStore.state.cart.totalNumItems().toString()
+        } else {
+            activity?.btnCartQuantity!!.visibility = View.GONE
+            activity?.btnTopCartQuantity!!.visibility = View.GONE
         }
     }
